csv_data_analysis_tool/requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
streamlit
python-dotenv
tabulate